apply focusmaster patch
This patch introduces a function similar to focusstack, but it focuses
on the next / previous client in the master area. This is useful with
layouts like deck, in which you might want to focus on a client in the
slave area, and then switch focus to a client in the master area.